feat(web): vanilla Three.js front served by the API (Milestone 2)
A zero-build web UI (toaster/web): ES modules + Three.js via CDN import-map, dark
theme. api.js decodes the {dtype,shape,data} arrays into TypedArrays; colors.js
computes point colours client-side from labels/grouping (so the wire carries no
colour buffers); viewer.js renders shader points with per-point colour + alpha
visibility, screen-space picking and rubber-band box select; app.js wires the
class/segmenter/groups panels, display modes, show/hide-solo and assign/undo to
the REST API. Served at / by the existing StaticFiles mount. Fixes selection to
serialize as int32 (JS has no int64 TypedArray).

+++ toaster/api/service.py
@@ -88,7 +88,8 @@ class AnnotationService:
             "snapshot": asdict(ctl.snapshot()),
             "labels": encode_array(ctl.label_array()),
             "grouping": encode_array(grouping) if grouping is not None else None,
-            "selection": encode_array(ctl.selection_indices()),
+            # int32 so it maps to a JS Int32Array (int64 has no plain TypedArray).
+            "selection": encode_array(ctl.selection_indices().astype(np.int32)),
         }
 
     # -- commands ---------------------------------------------------------
